+++ app/models/student_work.rb
@@ -4,6 +4,8 @@ class StudentWork < ApplicationRecord
   has_prefix_id :swrk
 
   has_many :student_criterion_levels, dependent: :destroy
+  has_many :feedback_items, as: :feedbackable
+
   belongs_to :assignment
   belongs_to :selected_document