+++ db/migrate/20250504032701_create_feedback_items.rb
@@ -1,0 +1,13 @@
+class CreateFeedbackItems < ActiveRecord::Migration[8.0]
+  def change
+    create_table :feedback_items do |t|
+      t.references :feedbackable, polymorphic: true, null: false
+      t.integer :item_type, null: false
+      t.string :title, null: false
+      t.text :description, null: false
+      t.text :evidence, null: false
+
+      t.timestamps
+    end
+  end
+end