+++ docs/changelog.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 - Added Level model with validations for criterion, title, description, and position.
 - Added StudentWork model with validations for assignment and selected document.
 - Added StudentCriterionLevel model with validations for student work, explanation, criterion, and level.
+- Added FeedbackItem model.
 
 ### Removed
 - Removed all classes from `application.html.erb` and `home/index.html.erb`, created backups of each.